Add X-Chekkit-Version header to check.php
This non-standard (as denoted by the X- prefix) is used during the authentication process to ensure that the correct version of Chekkit is being used. This field will need incrementing along with Chekkit.VERSION after every update. Outdated users will NOT be able to login if they are using an outdated version of Chekkit. If this header is not available, the check is ignored and authentication occurs as normal.